+++ src/components/FoodBox.js
@@ -1,0 +1,31 @@
+import { Card, Col, Button } from 'antd';
+
+function FoodBox({ food, removeFoodOnDelete }) {
+  const { name, calories, image, servings } = food;
+
+  return (
+    <Col>
+      <Card title={name} style={{ width: 230, height: 300, margin: 10 }}>
+        <img src={image} height={60} alt="food" />
+        <p>Calories: {calories}</p>
+        <p>Servings: {servings}</p>
+        <p>
+          <b>Total Calories: {calories * servings} </b> kcal
+        </p>
+        <Button
+          type="primary"
+          onClick={() => {
+            removeFoodOnDelete((arr) =>
+              arr.filter((dish) => dish.name !== name)
+            );
+          }}
+        >
+          {' '}
+          Delete{' '}
+        </Button>
+      </Card>
+    </Col>
+  );
+}
+
+export default FoodBox;
